Change coverage limit to 90% and fix/reorganize a few unit tests
