+++ setup.cfg
@@ -76,7 +76,7 @@ source = seismometer
 branch = True
 
 [coverage:report]
-fail_under = 95
+fail_under = 90
 omit =
     */plot/mpl/*
