+++ tests/controls/test_explore.py
@@ -13,6 +13,12 @@ from seismometer.controls.categorical_single_column import (
 from seismometer.data.performance import MetricGenerator
 from seismometer.table.analytics_table import AnalyticsTableOptionsWidget, ExploreBinaryModelAnalytics
 from seismometer.table.analytics_table_config import AnalyticsTableConfig
+from seismometer.table.fairness import (
+    BinaryClassifierMetricGenerator,
+    ExploreBinaryModelFairness,
+    FairnessOptionsWidget,
+    binary_metrics_fairness_table,
+)
 
 
 # region Test Base Classes
@@ -1534,6 +1540,68 @@ class TestCategoricalFeedbackSingleColumnOptionsWidget:
             model_options_widget=model_options_widget,
         )
         assert widget.model_options_widget == model_options_widget
+
+
+class TestExploreBinaryModelFairness:
+    def test_initialization_creates_components_correctly(self, monkeypatch):
+        sg_mock = Mock()
+        sg_mock.target_cols = ["readmit"]
+        sg_mock.output_list = ["risk_score"]
+        sg_mock.thresholds = [0.2, 0.5, 0.9]
+        sg_mock.available_cohort_groups = {"sex": ("M", "F")}
+        monkeypatch.setattr("seismometer.seismogram.Seismogram", lambda: sg_mock)
+
+        widget = ExploreBinaryModelFairness(rho=0.8)
+
+        # Ensure the correct metric generator is used
+        assert isinstance(widget.metric_generator, BinaryClassifierMetricGenerator)
+        assert widget.metric_generator.rho == 0.8
+
+        # Ensure the option widget is a FairnessOptionsWidget with the right model options
+        assert isinstance(widget.option_widget, FairnessOptionsWidget)
+        assert widget.option_widget.model_options.target == "readmit"
+        assert widget.option_widget.model_options.score == "risk_score"
+        assert widget.option_widget.model_options.thresholds["Score Threshold"] == 0.9
+
+        # Ensure correct plot function is set
+        assert widget.plot_function == binary_metrics_fairness_table
+
+    def test_generate_plot_args_returns_expected_values(self, monkeypatch):
+        # Mock Seismogram with sample model data
+        sg_mock = Mock()
+        sg_mock.target_cols = ["target_col"]
+        sg_mock.output_list = ["score_col"]
+        sg_mock.thresholds = [0.5]
+        sg_mock.available_cohort_groups = {"group": ("A", "B")}
+        monkeypatch.setattr("seismometer.seismogram.Seismogram", lambda: sg_mock)
+
+        # Instantiate widget
+        widget = ExploreBinaryModelFairness(rho=0.6)
+
+        # Simulate user selections
+        widget.option_widget.metric_list.value = ["Accuracy"]
+        widget.option_widget.cohort_list.value = {"group": ("A",)}
+        widget.option_widget.fairness_slider.value = 0.3
+
+        # Simulate model options
+        widget.option_widget.model_options.target_list.value = "target_col"
+        widget.option_widget.model_options.score_list.value = "score_col"
+        widget.option_widget.model_options.threshold_list.value = {"Score Threshold": 0.5}
+        widget.option_widget.model_options.per_context_checkbox.value = True
+
+        args, kwargs = widget.generate_plot_args()
+
+        # Assertions on returned values
+        assert args == (
+            widget.metric_generator,
+            ["Accuracy"],
+            {"group": ("A",)},
+            0.3,
+            "target_col",
+            "score_col",
+            0.5,
+        )
+        assert kwargs == {"per_context": True}
 
 
 # endregion
